Improve handling of Self
